Fix compilation under Beta4
