+++ Sources/BSWFoundation/Deferred/Deferred+Customizations.swift
@@ -153,31 +153,17 @@ extension Task.Result {
         }
     }    
 }
-
-import Combine
-
-@available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-public typealias CombineTask<T> = Publishers.Future<T, Swift.Error>
-
-@available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-public extension Task {
-    var future: CombineTask<Success> {
-        return .init { (promise) in
-            self.upon(DispatchQueue.any()) { (result) in
-                switch result {
-                case .success(let value):
-                    promise(.success(value))
-                case .failure(let error):
-                    promise(.failure(error))
-                }
-            }
+ 
+public extension Task.Result {
+    func extract() throws -> Success {
+        switch self {
+        case .success(let value):
+            return value
+        case .failure(let error):
+            throw error
         }
     }
-    var publisher: AnyPublisher<Success, Error> {
-        return future.eraseToAnyPublisher()
-    }
-}
-public extension Task.Result {
+    
     var swiftResult: Swift.Result<Success, Error> {
         switch self {
         case .failure(let error):
